Split commands into their own file
